requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.0
attrs==21.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
frozenlist==1.2.0
idna==2.10
multidict==5.2.0
pathlib==1.0.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.3.0
python-dotenv==0.17.1
requests==2.25.1
requests-toolbelt==0.9.1
typing-extensions==3.10.0.2
urllib3==1.26.4
yarl==1.7.0
